nunchuck converter
I used WiinUSoft to connect my Wii Remote and Nunchuck WIRELESSLY to my Windows Laptop. WiinUSoft emulates the Wii Remote as a Xbox 360 controller. This program takes the input from the singular joystick on the nunchuck and assigns it live to the right or left joystick, based on if the "C" button is pressed. It uses XInput-Python and Vgamepad. It is a little laggy, and I honestly don't know if it possible to optimize it further, but we will see.